Rename: pkg/netnsenter/netnsenter.go -> pkg/utils/nsenter/nsenter.go
This is no longer only for the net namespace.
